Extract native sidebar thread row
