+++ Sources/QuillCodeApp/QuillCodeSidebarThreadListView.swift
@@ -114,71 +114,3 @@ private struct QuillCodeSidebarThreadSectionView: View {
         .frame(maxWidth: .infinity, alignment: .leading)
     }
 }
-
-private struct QuillCodeSidebarThreadRowView: View {
-    var item: SidebarItemSurface
-    var isSelectionMode: Bool
-    var onSelectThread: (UUID) -> Void
-    var onThreadAction: (SidebarItemActionSurface) -> Void
-    var onCommand: (WorkspaceCommandSurface) -> Void
-
-    var body: some View {
-        HStack(spacing: 6) {
-            if isSelectionMode {
-                Button {
-                    toggleSelection()
-                } label: {
-                    Image(systemName: item.isBulkSelected ? "checkmark.circle.fill" : "circle")
-                        .font(.system(size: 16, weight: .semibold))
-                        .foregroundStyle(item.isBulkSelected ? QuillCodePalette.blue : QuillCodePalette.muted)
-                        .frame(width: QuillCodeMetrics.minimumHitTarget, height: QuillCodeMetrics.minimumHitTarget)
-                        .contentShape(Rectangle())
-                }
-                .buttonStyle(QuillCodePressableButtonStyle())
-                .accessibilityLabel(item.isBulkSelected ? "Deselect \(item.title)" : "Select \(item.title)")
-            }
-            Button {
-                if isSelectionMode {
-                    toggleSelection()
-                } else {
-                    onSelectThread(item.id)
-                }
-            } label: {
-                VStack(alignment: .leading, spacing: 4) {
-                    Text(item.title)
-                        .font(.callout.weight(.semibold))
-                        .lineLimit(1)
-                    Text(item.subtitle)
-                        .font(.caption)
-                        .foregroundStyle(QuillCodePalette.muted)
-                        .lineLimit(1)
-                }
-                .frame(maxWidth: .infinity, minHeight: QuillCodeMetrics.minimumHitTarget, alignment: .leading)
-            }
-            .buttonStyle(QuillCodePressableButtonStyle())
-
-            Menu {
-                ForEach(item.actions) { action in
-                    Button(role: action.kind == .delete ? .destructive : nil) {
-                        onThreadAction(action)
-                    } label: {
-                        Text(action.kind.title)
-                    }
-                }
-            } label: {
-                Image(systemName: "ellipsis")
-                    .frame(width: QuillCodeMetrics.minimumHitTarget, height: QuillCodeMetrics.minimumHitTarget)
-                    .foregroundStyle(QuillCodePalette.muted)
-                    .contentShape(Rectangle())
-            }
-            .buttonStyle(QuillCodePressableButtonStyle())
-        }
-        .padding(10)
-        .background(item.isSelected ? QuillCodePalette.selection : Color.clear)
-        .clipShape(RoundedRectangle(cornerRadius: 10, style: .continuous))
-    }
-
-    private func toggleSelection() {
-        onCommand(QuillCodeSidebarCommandAdapter.toggleSelectionCommand(for: item))
-    }
-}
