+++ Tests/QuillCodeParityTests/ParityGateTests.swift
@@ -1554,6 +1554,7 @@ final class ParityGateTests: QuillCodeParityTestCase {
         let adapterText = try Self.appSourceText(named: "QuillCodeSidebarCommandAdapter.swift")
         let sidebarText = try Self.appSourceText(named: "QuillCodeSidebarView.swift")
         let threadListText = try Self.appSourceText(named: "QuillCodeSidebarThreadListView.swift")
+        let threadRowText = try Self.appSourceText(named: "QuillCodeSidebarThreadRowView.swift")
         let projectListText = try Self.appSourceText(named: "QuillCodeProjectListView.swift")
         let htmlSidebarText = try Self.appSourceText(named: "WorkspaceHTMLSidebarRenderer.swift")
 
@@ -1569,7 +1570,8 @@ final class ParityGateTests: QuillCodeParityTestCase {
         XCTAssertTrue(sidebarText.contains("QuillCodeSidebarThreadListView"), "Native sidebar shell should delegate thread list and row rendering.")
         XCTAssertTrue(sidebarText.contains("QuillCodeProjectListView"), "Native sidebar shell should delegate project list and row rendering.")
         XCTAssertTrue(threadListText.contains("struct QuillCodeSidebarThreadListView"), "Thread list rendering should live in a focused native sidebar file.")
-        XCTAssertTrue(threadListText.contains("QuillCodeSidebarThreadRowView"), "Thread row rendering should live beside thread list rendering.")
+        XCTAssertTrue(threadListText.contains("QuillCodeSidebarThreadRowView"), "Thread list rendering should compose the focused thread row view.")
+        XCTAssertTrue(threadRowText.contains("struct QuillCodeSidebarThreadRowView"), "Thread row rendering should live in a focused native sidebar file.")
         XCTAssertTrue(projectListText.contains("struct QuillCodeProjectListView"), "Project list rendering should live in a focused native sidebar file.")
         XCTAssertTrue(projectListText.contains("QuillCodeProjectRowView"), "Project row rendering should live beside project list rendering.")
         XCTAssertTrue(sidebarText.contains("QuillCodeSidebarCommandPresentation.primaryCommandIDs"), "Native sidebar should consume shared primary command ordering.")
@@ -1578,13 +1580,14 @@ final class ParityGateTests: QuillCodeParityTestCase {
         XCTAssertTrue(sidebarText.contains("QuillCodeSidebarCommandPresentation.systemImage"), "Native sidebar should consume shared SF Symbols.")
         XCTAssertTrue(adapterText.contains("enum QuillCodeSidebarCommandAdapter"), "Sidebar command payload construction should live in a focused adapter.")
         XCTAssertTrue(sidebarText.contains("QuillCodeSidebarCommandAdapter.workspaceCommand") || threadListText.contains("QuillCodeSidebarCommandAdapter.workspaceCommand"), "Native sidebar should use the shared command adapter for bulk actions.")
-        XCTAssertTrue(threadListText.contains("QuillCodeSidebarCommandAdapter.toggleSelectionCommand"), "Native sidebar thread rows should use the shared command adapter for selection toggles.")
+        XCTAssertTrue(threadRowText.contains("QuillCodeSidebarCommandAdapter.toggleSelectionCommand"), "Native sidebar thread rows should use the shared command adapter for selection toggles.")
         XCTAssertTrue(htmlSidebarText.contains("renderPrimaryActions"), "HTML sidebar renderer should build primary sidebar actions through a helper.")
         XCTAssertTrue(htmlSidebarText.contains("renderUtilityActions"), "HTML sidebar renderer should build utility menu actions through a helper.")
         XCTAssertTrue(htmlSidebarText.contains("QuillCodeSidebarCommandPresentation.primaryCommandIDs"), "HTML sidebar renderer should consume shared primary command ordering.")
         XCTAssertTrue(htmlSidebarText.contains("QuillCodeSidebarCommandPresentation.visibleUtilityCommandGroups"), "HTML sidebar renderer should consume shared utility command groups.")
         XCTAssertTrue(htmlSidebarText.contains("QuillCodeSidebarCommandPresentation.htmlIconToken"), "HTML sidebar renderer should consume shared icon tokens.")
         XCTAssertFalse(sidebarText.contains("struct QuillCodeSidebarThreadRowView"), "Native sidebar shell should not own thread row rendering.")
+        XCTAssertFalse(threadListText.contains("private struct QuillCodeSidebarThreadRowView"), "Native sidebar thread list should not own thread row rendering.")
         XCTAssertFalse(sidebarText.contains("struct QuillCodeProjectRowView"), "Native sidebar shell should not own project row rendering.")
         XCTAssertFalse(sidebarText.contains("private func displayTitle"), "Native sidebar should not maintain a second label map.")
         XCTAssertFalse(sidebarText.contains("private func systemImage"), "Native sidebar should not maintain a second icon map.")
